v1.1.2

+++ lib/gridblock_var_replacer.php
@@ -2,7 +2,7 @@
 /*
 	Redaxo-Addon Gridblock
 	VAR-Replacer (REX-VARS und Widget-Buttons umschreiben)
-	v1.0.7
+	v1.1.2
 	by Falko Müller @ 2021-2022 (based on 0.1.0-dev von bloep)
 */
 
@@ -23,24 +23,28 @@ class rex_gridblock_var_replacer
 
     private static function replaceInput($colID, $moduleContent, $uID)
     {
-        $moduleContent = preg_replace('/REX_INPUT_VALUE\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][VALUE][$1]', $moduleContent);																	//UID + COLID an alle Inputs anhängen
-        $moduleContent =  preg_replace('/REX_INPUT_VALUE\['.$colID.'\]\['.$uID.'\]\[VALUE\]\[(\d+)\]\[(\d+)\]\[/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][VALUE][$1_MBLOCK][$2][', $moduleContent);			//MBLOCK-VARS zusätzlich aufbereiten
-
+        if ($colID !== NULL && $uID !== NULL && $moduleContent !== NULL):
+			$moduleContent = preg_replace('/REX_INPUT_VALUE\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][VALUE][$1]', $moduleContent);																	//UID + COLID an alle Inputs anhängen
+			$moduleContent =  preg_replace('/REX_INPUT_VALUE\['.$colID.'\]\['.$uID.'\]\[VALUE\]\[(\d+)\]\[(\d+)\]\[/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][VALUE][$1_MBLOCK][$2][', $moduleContent);			//MBLOCK-VARS zusätzlich aufbereiten
+		endif;
+		
         return $moduleContent;
     }
 	
 
     private static function replaceMedia($colID, $moduleContent, $uID)
     {
-		$moduleContent = preg_replace('/REX_INPUT_MEDIA\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][MEDIA][$1]', $moduleContent);
-        $moduleContent = preg_replace('/REXMedia\((\d+)/', 'REXMedia('.$colID.$uID.'00$1', $moduleContent);
-        $moduleContent = preg_replace('/id="REX_MEDIA_(\d+)"/', 'id="REX_MEDIA_'.$colID.$uID.'00$1"', $moduleContent);
-
-        //zusätzliche Ersetzung der MP-Buttons im Media-Widget
-        $moduleContent = preg_replace('/openREXMedia\(\'(\d+)\'/',				'openREXMedia(\''.$colID.$uID.'00$1\'', $moduleContent);
-        $moduleContent = preg_replace('/addREXMedia\(\'(\d+)\'/',				'addREXMedia(\''.$colID.$uID.'00$1\'', $moduleContent);
-        $moduleContent = preg_replace('/deleteREXMedia\(\'(\d+)\'/',			'deleteREXMedia(\''.$colID.$uID.'00$1\'', $moduleContent);
-        $moduleContent = preg_replace('/viewREXMedia\(\'(\d+)\'/',				'viewREXMedia(\''.$colID.$uID.'00$1\'', $moduleContent);
+		if ($colID !== NULL && $uID !== NULL && $moduleContent !== NULL):
+			$moduleContent = preg_replace('/REX_INPUT_MEDIA\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][MEDIA][$1]', $moduleContent);
+			$moduleContent = preg_replace('/REXMedia\((\d+)/', 'REXMedia('.$colID.$uID.'00$1', $moduleContent);
+			$moduleContent = preg_replace('/id="REX_MEDIA_(\d+)"/', 'id="REX_MEDIA_'.$colID.$uID.'00$1"', $moduleContent);
+	
+			//zusätzliche Ersetzung der MP-Buttons im Media-Widget
+			$moduleContent = preg_replace('/openREXMedia\(\'(\d+)\'/',				'openREXMedia(\''.$colID.$uID.'00$1\'', $moduleContent);
+			$moduleContent = preg_replace('/addREXMedia\(\'(\d+)\'/',				'addREXMedia(\''.$colID.$uID.'00$1\'', $moduleContent);
+			$moduleContent = preg_replace('/deleteREXMedia\(\'(\d+)\'/',			'deleteREXMedia(\''.$colID.$uID.'00$1\'', $moduleContent);
+			$moduleContent = preg_replace('/viewREXMedia\(\'(\d+)\'/',				'viewREXMedia(\''.$colID.$uID.'00$1\'', $moduleContent);
+		endif;
 		
         return $moduleContent;
     }
@@ -48,36 +52,40 @@ class rex_gridblock_var_replacer
 
     private static function replaceMediaList($colID, $moduleContent, $uID)
     {
-        $moduleContent = preg_replace('/REX_MEDIALIST_SELECT\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][MEDIALIST_SELECT][$1]', $moduleContent);
-        $moduleContent = preg_replace('/REX_INPUT_MEDIALIST\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][MEDIALIST][$1]', $moduleContent);
-        $moduleContent = preg_replace('/REXMedialist\((\d+)/', 'REXMedialist('.$colID.$uID.'00$1', $moduleContent);
-        $moduleContent = preg_replace('/id="REX_MEDIALIST_(\d+)"/', 'id="REX_MEDIALIST_'.$colID.$uID.'00$1"', $moduleContent);
-        $moduleContent = preg_replace('/id="REX_MEDIALIST_SELECT_(\d+)"/', 'id="REX_MEDIALIST_SELECT_'.$colID.$uID.'00$1"', $moduleContent);
-        
-        //zusätzliche Ersetzung der Buttons im Medialist-Widget
-        $moduleContent = preg_replace('/openREXMedialist\(\'(\d+)\'/',			'openREXMedialist(\''.$colID.$uID.'00$1\'', $moduleContent);
-        $moduleContent = preg_replace('/addREXMedialist\(\'(\d+)\'/',			'addREXMedialist(\''.$colID.$uID.'00$1\'', $moduleContent);
-        $moduleContent = preg_replace('/deleteREXMedialist\(\'(\d+)\'/',		'deleteREXMedialist(\''.$colID.$uID.'00$1\'', $moduleContent);
-        $moduleContent = preg_replace('/viewREXMedialist\(\'(\d+)\'/',			'viewREXMedialist(\''.$colID.$uID.'00$1\'', $moduleContent);
-        $moduleContent = preg_replace('/moveREXMedialist\(\'(\d+)\'/',			'moveREXMedialist(\''.$colID.$uID.'00$1\'', $moduleContent);
-		
+       	if ($colID !== NULL && $uID !== NULL && $moduleContent !== NULL):
+			$moduleContent = preg_replace('/REX_MEDIALIST_SELECT\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][MEDIALIST_SELECT][$1]', $moduleContent);
+			$moduleContent = preg_replace('/REX_INPUT_MEDIALIST\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][MEDIALIST][$1]', $moduleContent);
+			$moduleContent = preg_replace('/REXMedialist\((\d+)/', 'REXMedialist('.$colID.$uID.'00$1', $moduleContent);
+			$moduleContent = preg_replace('/id="REX_MEDIALIST_(\d+)"/', 'id="REX_MEDIALIST_'.$colID.$uID.'00$1"', $moduleContent);
+			$moduleContent = preg_replace('/id="REX_MEDIALIST_SELECT_(\d+)"/', 'id="REX_MEDIALIST_SELECT_'.$colID.$uID.'00$1"', $moduleContent);
+			
+			//zusätzliche Ersetzung der Buttons im Medialist-Widget
+			$moduleContent = preg_replace('/openREXMedialist\(\'(\d+)\'/',			'openREXMedialist(\''.$colID.$uID.'00$1\'', $moduleContent);
+			$moduleContent = preg_replace('/addREXMedialist\(\'(\d+)\'/',			'addREXMedialist(\''.$colID.$uID.'00$1\'', $moduleContent);
+			$moduleContent = preg_replace('/deleteREXMedialist\(\'(\d+)\'/',		'deleteREXMedialist(\''.$colID.$uID.'00$1\'', $moduleContent);
+			$moduleContent = preg_replace('/viewREXMedialist\(\'(\d+)\'/',			'viewREXMedialist(\''.$colID.$uID.'00$1\'', $moduleContent);
+			$moduleContent = preg_replace('/moveREXMedialist\(\'(\d+)\'/',			'moveREXMedialist(\''.$colID.$uID.'00$1\'', $moduleContent);
+		endif;
+				
         return $moduleContent;
     }
 	
 
     private static function replaceLink($colID, $moduleContent, $uID)
     {
-        $moduleContent = preg_replace('/REX_LINK_NAME\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][LINK_NAME][$1]', $moduleContent);
-        $moduleContent = preg_replace('/id="REX_LINK_(\d+)_NAME"/', 'id="REX_LINK_'.$colID.$uID.'00$1_NAME"', $moduleContent);
-
-        $moduleContent = preg_replace('/REX_INPUT_LINK\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][LINK][$1]', $moduleContent);
-        $moduleContent = preg_replace('/openLinkMap\(\'REX_LINK_(\d+)\'/', 'openLinkMap(\'REX_LINK_'.$colID.$uID.'00$1\'', $moduleContent);
-        $moduleContent = preg_replace('/REXLink\((\d+)/', 'REXLink('.$colID.$uID.'00$1', $moduleContent);
-        $moduleContent = preg_replace('/id="REX_LINK_(\d+)"/', 'id="REX_LINK_'.$colID.$uID.'00$1"', $moduleContent);
-
-        //zusätzliche Ersetzung der Buttons im Link-Widget
-        $moduleContent = preg_replace('/openLinkMap\(\'REX_LINK_(\d+)\'/',		'openLinkMap(\'REX_LINK_'.$colID.$uID.'00$1\'', $moduleContent);
-        $moduleContent = preg_replace('/deleteREXLink\(\'(\d+)\'/',				'deleteREXLink(\''.$colID.$uID.'00$1\'', $moduleContent);
+		if ($colID !== NULL && $uID !== NULL && $moduleContent !== NULL):
+			$moduleContent = preg_replace('/REX_LINK_NAME\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][LINK_NAME][$1]', $moduleContent);
+			$moduleContent = preg_replace('/id="REX_LINK_(\d+)_NAME"/', 'id="REX_LINK_'.$colID.$uID.'00$1_NAME"', $moduleContent);
+	
+			$moduleContent = preg_replace('/REX_INPUT_LINK\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][LINK][$1]', $moduleContent);
+			$moduleContent = preg_replace('/openLinkMap\(\'REX_LINK_(\d+)\'/', 'openLinkMap(\'REX_LINK_'.$colID.$uID.'00$1\'', $moduleContent);
+			$moduleContent = preg_replace('/REXLink\((\d+)/', 'REXLink('.$colID.$uID.'00$1', $moduleContent);
+			$moduleContent = preg_replace('/id="REX_LINK_(\d+)"/', 'id="REX_LINK_'.$colID.$uID.'00$1"', $moduleContent);
+	
+			//zusätzliche Ersetzung der Buttons im Link-Widget
+			$moduleContent = preg_replace('/openLinkMap\(\'REX_LINK_(\d+)\'/',		'openLinkMap(\'REX_LINK_'.$colID.$uID.'00$1\'', $moduleContent);
+			$moduleContent = preg_replace('/deleteREXLink\(\'(\d+)\'/',				'deleteREXLink(\''.$colID.$uID.'00$1\'', $moduleContent);
+		endif;
 		
         return $moduleContent;
     }
@@ -85,16 +93,18 @@ class rex_gridblock_var_replacer
 
     private static function replaceLinkList($colID, $moduleContent, $uID)
     {
-        $moduleContent = preg_replace('/REX_LINKLIST_SELECT\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][LINKLIST_SELECT][$1]', $moduleContent);
-        $moduleContent = preg_replace('/REX_INPUT_LINKLIST\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][LINKLIST][$1]', $moduleContent);
-        $moduleContent = preg_replace('/REXLinklist\((\d+)/', 'REXLinklist('.$colID.$uID.'00$1', $moduleContent);
-        $moduleContent = preg_replace('/id="REX_LINKLIST_(\d+)"/', 'id="REX_LINKLIST_'.$colID.$uID.'00$1"', $moduleContent);
-        $moduleContent = preg_replace('/id="REX_LINKLIST_SELECT_(\d+)"/', 'id="REX_LINKLIST_SELECT_'.$colID.$uID.'00$1"', $moduleContent);
-
-        //zusätzliche Ersetzung der Buttons im Linklist-Widget
-        $moduleContent = preg_replace('/openREXLinklist\(\'(\d+)\'/',			'openREXLinklist(\''.$colID.$uID.'00$1\'', $moduleContent);
-        $moduleContent = preg_replace('/deleteREXLinklist\(\'(\d+)\'/',			'deleteREXLinklist(\''.$colID.$uID.'00$1\'', $moduleContent);
-        $moduleContent = preg_replace('/moveREXLinklist\(\'(\d+)\'/',			'moveREXMedialist(\''.$colID.$uID.'00$1\'', $moduleContent);
+		if ($colID !== NULL && $uID !== NULL && $moduleContent !== NULL):
+			$moduleContent = preg_replace('/REX_LINKLIST_SELECT\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][LINKLIST_SELECT][$1]', $moduleContent);
+			$moduleContent = preg_replace('/REX_INPUT_LINKLIST\[(\d+)\]/', 'REX_INPUT_VALUE['.$colID.']['.$uID.'][LINKLIST][$1]', $moduleContent);
+			$moduleContent = preg_replace('/REXLinklist\((\d+)/', 'REXLinklist('.$colID.$uID.'00$1', $moduleContent);
+			$moduleContent = preg_replace('/id="REX_LINKLIST_(\d+)"/', 'id="REX_LINKLIST_'.$colID.$uID.'00$1"', $moduleContent);
+			$moduleContent = preg_replace('/id="REX_LINKLIST_SELECT_(\d+)"/', 'id="REX_LINKLIST_SELECT_'.$colID.$uID.'00$1"', $moduleContent);
+	
+			//zusätzliche Ersetzung der Buttons im Linklist-Widget
+			$moduleContent = preg_replace('/openREXLinklist\(\'(\d+)\'/',			'openREXLinklist(\''.$colID.$uID.'00$1\'', $moduleContent);
+			$moduleContent = preg_replace('/deleteREXLinklist\(\'(\d+)\'/',			'deleteREXLinklist(\''.$colID.$uID.'00$1\'', $moduleContent);
+			$moduleContent = preg_replace('/moveREXLinklist\(\'(\d+)\'/',			'moveREXMedialist(\''.$colID.$uID.'00$1\'', $moduleContent);
+		endif;
 		
         return $moduleContent;
     }
@@ -103,19 +113,40 @@ class rex_gridblock_var_replacer
     public static function replaceModuleVars($moduleContent, $rexVars)
     {
 		//REX-Vars
-		$moduleContent = preg_replace('/REX_CTYPE_ID/', 				@$rexVars['ctypeID'], $moduleContent);				//wird nicht unterstützt
-        $moduleContent = preg_replace('/REX_SLICE_ID/', 				@$rexVars['sliceID'], $moduleContent);
-		$moduleContent = preg_replace('/REX_MODULE_ID/',				@$rexVars['moduleID'], $moduleContent);
-		$moduleContent = preg_replace('/REX_MODULE_KEY/', 				@$rexVars['moduleKEY'], $moduleContent);
-		$moduleContent = preg_replace('/REX_CLANG_ID/', 				@$rexVars['clangID'], $moduleContent);
-		$moduleContent = preg_replace('/REX_ARTICLE_ID/', 				@$rexVars['artID'], $moduleContent);
-		$moduleContent = preg_replace('/REX_TEMPLATE_ID/', 				@$rexVars['tmplID'], $moduleContent);
+		$tmp = @$rexVars['ctypeID']; 		$tmp = ($tmp !== NULL) ? $tmp : '';
+		$moduleContent = preg_replace('/REX_CTYPE_ID/', 	$tmp, $moduleContent);
+			
+		$tmp = @$rexVars['sliceID']; 		$tmp = ($tmp !== NULL) ? $tmp : '';
+        $moduleContent = preg_replace('/REX_SLICE_ID/', 	$tmp, $moduleContent);
+
+		$tmp = @$rexVars['moduleID']; 		$tmp = ($tmp !== NULL) ? $tmp : '';
+		$moduleContent = preg_replace('/REX_MODULE_ID/',	$tmp, $moduleContent);
+
+		$tmp = @$rexVars['moduleKEY']; 		$tmp = ($tmp !== NULL) ? $tmp : '';
+		$moduleContent = preg_replace('/REX_MODULE_KEY/',	$tmp, $moduleContent);
+		
+		$tmp = @$rexVars['clangID']; 		$tmp = ($tmp !== NULL) ? $tmp : '';
+		$moduleContent = preg_replace('/REX_CLANG_ID/', 	$tmp, $moduleContent);
+		
+		$tmp = @$rexVars['artID']; 			$tmp = ($tmp !== NULL) ? $tmp : '';
+		$moduleContent = preg_replace('/REX_ARTICLE_ID/',	$tmp, $moduleContent);
+		
+		$tmp = @$rexVars['tmplID']; 		$tmp = ($tmp !== NULL) ? $tmp : '';
+		$moduleContent = preg_replace('/REX_TEMPLATE_ID/',	$tmp, $moduleContent);
+		
 		
 		//GRID-Vars
-		$moduleContent = preg_replace('/REX_GRID_TEMPLATE_ID/', 		@$rexVars['grid_tmplID'], $moduleContent);
-		$moduleContent = preg_replace('/REX_GRID_TEMPLATE_PREVIEW/', 	@$rexVars['grid_tmplPREV'], $moduleContent);
-		$moduleContent = preg_replace('/REX_GRID_TEMPLATE_COLUMNS/', 	@$rexVars['grid_tmplCOLS'], $moduleContent);
-		$moduleContent = preg_replace('/REX_GRID_COLUMN_NUMBER/', 		@$rexVars['grid_colNR'], $moduleContent);
+		$tmp = @$rexVars['grid_tmplID']; 	$tmp = ($tmp !== NULL) ? $tmp : '';
+		$moduleContent = preg_replace('/REX_GRID_TEMPLATE_ID/', 	 $tmp, $moduleContent);
+		
+		$tmp = @$rexVars['grid_tmplPREV']; 	$tmp = ($tmp !== NULL) ? $tmp : '';
+		$moduleContent = preg_replace('/REX_GRID_TEMPLATE_PREVIEW/', $tmp, $moduleContent);
+		
+		$tmp = @$rexVars['grid_tmplCOLS']; 	$tmp = ($tmp !== NULL) ? $tmp : '';
+		$moduleContent = preg_replace('/REX_GRID_TEMPLATE_COLUMNS/', $tmp, $moduleContent);
+		
+		$tmp = @$rexVars['grid_colNR']; 	$tmp = ($tmp !== NULL) ? $tmp : '';
+		$moduleContent = preg_replace('/REX_GRID_COLUMN_NUMBER/', 	 $tmp, $moduleContent);
 		
         return $moduleContent;
     }
